+++ netlify/functions/confirmar-presente.js
@@ -1,17 +1,16 @@
 const fs = require('fs').promises;
 const { getWritablePath } = require('./lib/fileHelper');
+const { obterListaProdutos, salvarListaProdutos } = require('./lib/produtos');
 const API_URL = process.env.API_URL;
 
-const file = getWritablePath('controle-de-produto.json');
 const mensagensFile = getWritablePath('mensagens.json');
 
 async function carregarProdutos() {
-  const data = await fs.readFile(file, 'utf8');
-  return JSON.parse(data);
+  return await obterListaProdutos();
 }
 
 async function salvarProdutos(lista) {
-  await fs.writeFile(file, JSON.stringify(lista, null, 2));
+  await salvarListaProdutos(lista);
 }
 
 async function carregarMensagens() {
